Move member to this service instead of being a general extension
