+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Extensions/SourceTextExtensions_SharedWithCodeStyle.cs
@@ -70,26 +70,5 @@ namespace Microsoft.CodeAnalysis.Shared.Extensions
 
         public static bool AreOnSameLine(this SourceText text, int pos1, int pos2)
             => text.Lines.IndexOf(pos1) == text.Lines.IndexOf(pos2);
-
-        public static bool IsBeforeOrAfterNodeOnSameLine(
-            this SourceText text, SyntaxNode root, SyntaxNode member, int position)
-        {
-            var token = root.FindToken(position);
-            if (token == member.GetFirstToken() &&
-                position <= token.SpanStart &&
-                text.AreOnSameLine(position, token.SpanStart))
-            {
-                return true;
-            }
-
-            if (token == member.GetLastToken() &&
-                position >= token.Span.End &&
-                text.AreOnSameLine(position, token.Span.End))
-            {
-                return true;
-            }
-
-            return false;
-        }
     }
 }
